member 도메인 service, controller 리팩토링, rq 도입

+++ backend/src/main/java/com/back/p67260811/domain/post/post/controller/ApiV1PostController.java
@@ -7,6 +7,7 @@ import com.back.p67260811.domain.post.post.entity.Post;
 import com.back.p67260811.domain.post.post.service.PostService;
 import com.back.p67260811.global.dto.RsData;
 import com.back.p67260811.global.exception.ServiceException;
+import com.back.p67260811.global.rq.Rq;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
@@ -24,6 +25,7 @@ public class ApiV1PostController {
 
     private final PostService postService;
     private final MemberService memberService;
+    private final Rq rq;
 
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
     public List<PostDto> list() {
@@ -62,16 +64,11 @@ public class ApiV1PostController {
     @PostMapping
     @Transactional
     public RsData<PostDto> write(
-            @Valid @RequestBody PostWriteReqBody reqBody,
-            @RequestHeader("Authorization") String authorization
+            @Valid @RequestBody PostWriteReqBody reqBody
     ) {
-
-        String apiKey = authorization.substring(7);
-        Member actor = memberService.findByApiKey(apiKey).orElseThrow(
-                () -> new ServiceException("401-1", "API Key가 유효하지 않습니다.")
-        );
-
+        Member actor = rq.getActor();
         Post post = postService.write(actor, reqBody.title, reqBody.content);
+
         return new RsData<>(
                 "201-1",
                 "%d번 글이 성공적으로 등록되었습니다".formatted(post.getId()),
@@ -84,7 +81,7 @@ public class ApiV1PostController {
             @Size(min = 2, max = 10, message = "제목은 2글자 이상 10글자 이하로 작성해주세요.")
             @NotBlank(message = "제목을 입력해주세요.")
             String title,
-            @Size(min = 2, max = 10, message = "내용은 2글자 이상 10글자 이하로 작성해주세요.")
+            @Size(min = 2, message = "내용은 2글자 이상 작성해주세요.")
             @NotBlank(message = "내용을 입력해주세요.")
             String content
     ) {
@@ -96,7 +93,13 @@ public class ApiV1PostController {
             @PathVariable int id,
             @Valid @RequestBody PostModifyReqBody reqBody
     ) {
+        Member actor = rq.getActor();
         Post post = postService.findById(id).get();
+
+        if (actor.getId() != post.getAuthor().getId()) {
+            throw new ServiceException("403-1", "수정 권한이 없습니다.");
+        }
+
         postService.modify(post, reqBody.title, reqBody.content);
 
         return new RsData<>(
@@ -109,6 +112,13 @@ public class ApiV1PostController {
     public RsData<Void> delete(
             @PathVariable int id
     ) {
+        Member actor = rq.getActor();
+        Post post = postService.findById(id).get();
+
+        if (!actor.equals(post.getAuthor())) {
+            throw new ServiceException("403-1", "삭제 권한이 없습니다.");
+        }
+
         postService.delete(id);
 
         return new RsData<>(
